Successfully seeding 10M records to PG from csv
