+++ .env
@@ -1,4 +1,4 @@
-POSTGRES_DB=songs
+POSTGRES_DEV_DB=songs
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 POSTGRES_PORT=5432
