+++ knexfile.js
@@ -6,7 +6,7 @@ module.exports = {
   development: {
     client: 'pg',
     connection: {
-      database: process.env.POSTGRES_DB,
+      database: process.env.POSTGRES_DEV_DB,
       user: process.env.POSTGRES_USER,
       password: process.env.POSTGRES_PASSWORD,
     },
